+++ packages/orq-rc/RELEASES.md
@@ -39,3 +39,13 @@ Based on:
 - [typescript v1.41.0-rc.15] packages/orq-rc
 ### Releases
 - [NPM v1.41.0-rc.15] https://www.npmjs.com/package/orq-poc-typescript-multi-env-version/v/1.41.0-rc.15 - packages/orq-rc
+
+## 2024-11-25 06:59:56
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.444.4 (2.461.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.42.0-rc.1] packages/orq-rc
+### Releases
+- [NPM v1.42.0-rc.1] https://www.npmjs.com/package/orq-poc-typescript-multi-env-version/v/1.42.0-rc.1 - packages/orq-rc